% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slab_functions.R
\name{add_topo_stack}
\alias{add_topo_stack}
\title{Title}
\usage{
add_topo_stack(slab, fact = 1, extend.lims = c(-1, 1, -1, 1),
  etopo = "/Volumes/data/data/global/topography/etopo/ETOPO1_Bed_g_geotiff.tif")
}
\arguments{
\item{etopo}{}
}
\description{
Title
}
