% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slab_functions.R
\name{assemble_slab}
\alias{assemble_slab}
\title{Title}
\usage{
assemble_slab(
  slab = NA,
  slab.name = "izu",
  extend.lims = c(-1, 1, -1, 1),
  slab.dir = "/Volumes/data/data/global/slabs/slab1/allslabs/",
  etopo = "~/Dropbox/data/global/topography/etopo/ETOPO1_Bed_g_geotiff.tif",
  slab2 = T,
  fact = 1,
  detail = "h",
  borders = F,
  plates = F,
  rivers = T,
  coasts = T,
  ppp = F,
  simplify = F,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
