% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_funcs.R
\name{azdip2xyz}
\alias{azdip2xyz}
\title{azdip2xyz}
\usage{
azdip2xyz(az = 20, dip = 60, orth = F, str = F, ver = F)
}
\arguments{
\item{orth}{logical default is FALSE, specifie if dip}

\item{str}{logical default is FALSE, specify if strike tren is provide instead of azimuth, assuems right hand rule (aziumth is striike +90 modulo 360)}
}
\value{
xyz
}
\description{
azdip2xyz
}
\examples{
azdip(az=20, dip=60, orth=F, str=F)


}
