% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{build_raster_stack}
\alias{build_raster_stack}
\title{Title}
\usage{
build_raster_stack(ras, shade = c(300, 10, 320), shade.marine = c(10, 20,
  320), thresh = 0, topo = FALSE)
}
\arguments{
\item{ras}{raw raster.}

\item{shade}{length 3 vector, vertical.exag, sun.elevation, sun.azimumuth.}

\item{thresh}{length 1 c(lower) or two c(lower, upper) vector for threshold settings.}

\item{topo}{}

\item{.marine}{shade length 3 vector, vertical.exag, sun.elevation, sun.azimuth.}

\item{topo}{only clauclates toposhade}
}
\value{
raster stack of 4 layers.
}
\description{
Title
}
\examples{
create a raster stack with elements, topo - raw raster .
topo.thresh - thesholded  raster according to theresh params
shade - hill shaded  raw raster using shade params
shade.marine - hill shaded  raster < thrsh[1] using shade.marine params
shade.theesh - hill shaded  raster > thrsh[1] using shade.marine params
ras<-build_raster_stack(bt.1d.median  \%>\% set_which(rmax=1500), shade=c(600,5,320))
}
