% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{convertLims}
\alias{convertLims}
\title{Title}
\usage{
convertLims(limsx, limsy, from = CRS("+init=epsg:4326"),
  to = CRS("+proj=laea +lat_0=-45.7 +lon_0=132.9 +ellps=WGS84"))
}
\arguments{
\item{to}{}
}
\description{
Title
}
