% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{DEMderiv}
\alias{DEMderiv}
\title{Title}
\usage{
DEMderiv(data, attr = "plan.curvature", method = "evans")
}
\arguments{
\item{method}{}

\item{neighb}{}
}
\value{







}
\description{
Title
}
\examples{
Title

Title

Title

}
