% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_funcs.R
\name{df2spline}
\alias{df2spline}
\title{Computes a spatile lines object froma  data frame with long lat and id columns
afatped from https://rpubs.com/walkerke/points_to_line}
\usage{
df2spline(df, proj.string = "+init=epsg:4326")
}
\arguments{
\item{data}{}
}
\value{
SpatialLines
}
\description{
Computes a spatile lines object froma  data frame with long lat and id columns
afatped from https://rpubs.com/walkerke/points_to_line
}
\examples{
 plates= get_plates()
 df_to spline(plates)

}
