% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_funcs.R
\name{get_ghss}
\alias{get_ghss}
\title{Title}
\usage{
get_ghss(fn = get_gshhg_file(), xlim = c(138, 152), ylim = c(-41.5, -33),
  factor = 0.25, level = 1, no.clip = F, properly = T,
  checkPolygons = F, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
