% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_funcs.R
\name{get_ghss1}
\alias{get_ghss1}
\title{Title}
\usage{
get_ghss1(gshhg = "~/data/global/coastlines/gshhg-bin-2/", xlim = c(138,
  152), ylim = c(-41.5, -33), grid = c(4, 2), detail = "i",
  factor = 0.25, projection = NULL, level = 1, no.clip = F,
  properly = T, checkPolygons = F, rivers = F, borders = F, shift = F,
  ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
