% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_funcs.R
\name{get_gshhg_file}
\alias{get_gshhg_file}
\title{Title}
\usage{
get_gshhg_file(
  dir = "~/data/global/coastlines/gshhg-shp-2/",
  type = "GSHHS",
  resolution = "i",
  level = 1
)
}
\arguments{
\item{level}{}
}
\value{

}
\description{
Title
}
