% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_poly}
\alias{get_poly}
\title{Title}
\usage{
get_poly(long = -100, lat = 15, length = 5, width = 2, angle = 15,
  owin = T, clockwise = F, left = 1, right = 1, top = 1, bottom = 1)
}
\arguments{
\item{bottom}{}
}
\description{
Title
}
