% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_funcs.R
\name{get_wdbii_file}
\alias{get_wdbii_file}
\title{Title}
\usage{
get_wdbii_file(
  dir = "~/data/global/coastlines/gshhg-shp-2/",
  type = "WDBII",
  resolution = "i",
  level = 1,
  river = F,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
