% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graticules.R
\name{graticule_longs}
\alias{graticule_longs}
\title{Title}
\usage{
graticule_longs(vec, inc = 0.1, proj4string = CRS("+init=epsg:4326"))
}
\arguments{
\item{proj4string}{}
}
\value{

}
\description{
Title
}
