% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_funcs.R
\name{lines2angle}
\alias{lines2angle}
\title{the angle in dgrees between two lines}
\usage{
lines2angle(
  l1 = azdip2xyz(),
  l2 = azdip2xyz(dip = -30, ver = FALSE),
  ver = FALSE,
  sig = 1
)
}
\arguments{
\item{l1}{line 1 as c(x,y,z) or c(az, dip)}

\item{l2}{line2 as c(x,y,z) or c(az, dip)}

\item{ver}{verbose}

\item{sig}{significnace in round resauklt (default=1)}
}
\value{
numeric
}
\description{
the angle in dgrees between two lines
}
\examples{
l1 <- c(330,20)
l2 <- c(220, 49)
lines2angle(l1,l2)

}
