% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slab_functions.R
\name{load_slab}
\alias{load_slab}
\title{Title}
\usage{
load_slab(slab.name = "izu",
  slab.dir = "/Users/msandifo/Dropbox/msandifo/documents/programming/r/2017/ecuador/data/slab1/allslabs/")
}
\arguments{
\item{slab.dir}{}
}
\description{
Title
}
