% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slab_functions.R
\name{merge_slabs}
\alias{merge_slabs}
\title{Title}
\usage{
merge_slabs(slabs = list(mex, sam), extend.lims = c(0, 0, 0, 0),
  slab.name = "mex.sam")
}
\arguments{
\item{slab.name}{}
}
\description{
Title
}
