% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cv_model.R
\name{plot.cv_model}
\alias{plot.cv_model}
\title{Plot Cross-Validation Coefficients}
\usage{
\method{plot}{cv_model}(result_model)
}
\arguments{
\item{result_model}{A model object returned by cv.glmnet containing cross-validation results.
This object should include cvm (cross-validation error mean), cvsd (cross-validation error standard deviation),
lambda_range (range of lambda values), and lambda (optimal lambda value).}
}
\value{
This function does not return a value. It generates a plot.
}
\description{
This function plots the cross-validation error against the log of lambda values
for a given glmnet model. It visualizes the performance of the model across different levels
of regularization.
}
\examples{
# Same data generate process as slab_shrink
set.seed(123)
n <- 100 # Number of observations
p <- 50  # Number of variables
test_x <- matrix(rnorm(n * p), n, p)
beta <- matrix(rnorm(p + 1), p + 1, 1)
test_y <- cbind(1, test_x) \%*\% beta + rnorm(n, sd = 0.5)

# Fit a model using a fictional slab_shrink function
result <- slab_shrink(test_x, test_y) # Estimation works with just x and y

# Plot the model using default settings
plot.cv_model(result)

}
