% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{plot_axes}
\alias{plot_axes}
\title{Title}
\usage{
plot_axes(ras, ax.ext = 0.015, n.minor = 10, lwd = 2, tick.sep = 5,
  lwd.ticks = 1.5, lwd.ticks.minor = 1.5, major = 0.01,
  minor = 0.004, clip = T, minute = TRUE, sec = FALSE, trad = NA,
  cex = 2.5, add = F, grid = T, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
