% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_slab}
\alias{plot_slab}
\title{Main slab  plotting fucntion}
\usage{
plot_slab(slab, slab.col = stretch_brewer(pal = "Spectral", stretch = 0.5),
  col = NA, out.size = 1200, stretch.ras = c(0.15, 0.5),
  stretch.plan = c(0.6, 0.8), alpha.plan = 1, mgp = c(0.5, 0, 0),
  smallplot = c(0.35, 0.65, 0.04, 0.055), cex = 2, legcex = 2,
  alpha = 0.75, contours = c(-10, -30, -70, -(seq(100, 600, 100))),
  ax.ext = c(0.015, 0.015), hf = FALSE, ehb = TRUE, wsm = TRUE,
  cmt = TRUE, wscale = 0.5, wlwd = 2, Pax = T, Tax = F, depth = NA,
  ends = T, shade = T, plancurv = F, trad = T, shallow = F, dip = T,
  ...)
}
\arguments{
\item{...}{}
}
\description{
Main slab  plotting fucntion
}
