% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_wsm}
\alias{plot_wsm}
\title{Plot world stree map data}
\usage{
plot_wsm(
  wsm = NA,
  scale = 1,
  lines = T,
  limsx = c(-180, 180),
  limsy = c(-90, 90),
  lwd = 1,
  lty = 1,
  shade = TRUE,
  Tax = F,
  Pax = F,
  ends = TRUE
)
}
\arguments{
\item{ends}{}
}
\value{

}
\description{
Plot world stree map data
}
