% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{pnts2ras}
\alias{pnts2ras}
\title{Title}
\usage{
pnts2ras(cad.sp.1, res = c(8, 16, 32, 64), init.space = c(0.5, 0.5),
  method = "", asis = FALSE)
}
\arguments{
\item{asis}{}
}
\description{
Title
}
