% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slab_functions.R
\name{prepare_ehb}
\alias{prepare_ehb}
\title{Title}
\usage{
prepare_ehb(
  dirname = "/Volumes/data/data/global/quakes/ehb19",
  save = T,
  ppp = T
)
}
\arguments{
\item{ppp}{}
}
\value{

}
\description{
Title
}
