% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slab_functions.R
\name{ras_plancurv}
\alias{ras_plancurv}
\title{Title}
\usage{
ras_plancurv(ras, method = "evans", fact = 1, foc = 25,
  stretch.plan = c(0.25, 0.75))
}
\arguments{
\item{stretch.plan}{}
}
\description{
Title
}
