% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slab_functions.R
\name{read_cmt}
\alias{read_cmt}
\title{Title}
\usage{
read_cmt(
  fname = "~/Dropbox/data/global/quakes/cmt/cmt1.RData",
  limsx = c(-180, 180),
  limsy = c(-90, 90),
  ppp = T,
  simplify = F
)
}
\arguments{
\item{simplify}{}
}
\value{

}
\description{
Title
}
