% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slab_functions.R
\name{read_ehb}
\alias{read_ehb}
\title{Title}
\usage{
read_ehb(ehb = NA, limsx = c(-180, 180), limsy = c(-90, 90), ppp = TRUE,
  to = NA, fname = "/Volumes/data/data/global/quakes/ehb19/ehb.Rdata")
}
\arguments{
\item{fname}{}
}
\description{
Title
}
