% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slab_functions.R
\name{read_hf}
\alias{read_hf}
\title{Title}
\usage{
read_hf(ghf.file = "~/Dropbox/data/global/heatflow/Global2010.csv",
  limsx = c(-360, 360), limsy = c(-90, 90), ppp = TRUE)
}
\arguments{
\item{ppp}{}
}
\description{
Title
}
