% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slab_functions.R
\name{read_vo}
\alias{read_vo}
\title{Title}
\usage{
read_vo(vfile = "~/Dropbox/data/global/volcanoes/volcano1.csv",
  limsx = c(-180, 180), limsy = c(-90, 90), ppp = TRUE, to = NA)
}
\arguments{
\item{to}{}
}
\description{
Title
}
