% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slab_functions.R
\name{read_vo1}
\alias{read_vo1}
\title{Title}
\usage{
read_vo1(vfile = "~/Dropbox/data/global/volcanoes/", limsx = c(-180, 180),
  limsy = c(-90, 90), ppp = TRUE, to = NA)
}
\arguments{
\item{to}{}
}
\description{
Title
}
