% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{trad_axes_poly_back}
\alias{trad_axes_poly_back}
\title{Title}
\usage{
trad_axes_poly_back(
  e.ras,
  axis.width = 0.007,
  width = 2,
  crs = CRS("+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0"),
  nseq = 51,
  extend = 0.002
)
}
\arguments{
\item{extend}{}
}
\value{

}
\description{
Title
}
