% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slab_functions.R
\name{wsm_ppp}
\alias{wsm_ppp}
\title{Title}
\usage{
wsm_ppp(wsm = NA, fname = "~/Dropbox/data/global/wsm/wsm2016.csv",
  sep = ",", limsx = c(-180, 180), limsy = c(-90, 90), ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
