% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slab_functions.R
\name{assemble_slab}
\alias{assemble_slab}
\title{Title}
\usage{
assemble_slab(slab = NA, slab.name = "izu", extend.lims = c(-1, 1,
  -1, 1), slab.dir = "/Volumes/data/data/global/slabs/slab1/allslabs/",
  etopo = "~/Dropbox/data/global/topography/etopo/ETOPO1_Bed_g_geotiff.tif",
  slab2 = F, fact = 1, detail = "h", borders = T, plates = T,
  rivers = T, coasts = T, ppp = T, simplify = F, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
