% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slab_functions.R
\name{load_slab}
\alias{load_slab}
\title{Title}
\usage{
load_slab(
  slab.name = "izu",
  slab.dir = "/Volumes/data/data/global/slabs/slab1/slabs"
)
}
\arguments{
\item{slab.dir}{}
}
\value{

}
\description{
Title
}
